+++ MainGUI.py
@@ -13,10 +13,10 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
     def __init__(self, parent=None):
         super(MainGUI, self).__init__()
         self.setupUi(self)
-        self.show()
 
         self.disable_buttons()
         self.Config = Parameters.Parameters({})
+        self.show()
 
         self.LoadConfigButton.clicked.connect(self.load_configures)
         self.ImageSetBrowseButton.clicked.connect(lambda: self.ShowFileDialog(1))
@@ -26,6 +26,9 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
         self.SaveResButton.clicked.connect(self.save_results)
         self.DelModelButton.clicked.connect(self.delete_model)
         self.StartPredictButton.clicked.connect(self.picture_predict)
+        self.WrongButton.clicked.connect(lambda: self.wrong_prediction(self.feed_backs))
+        self.RightButton.clicked.connect(lambda: self.right_prediction(self.feed_backs))
+
 
     def closeEvent(self, *args, **kwargs):
         # 关闭窗口时保存配置
@@ -40,6 +43,7 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
         except:
             QtWidgets.QMessageBox.warning(self, 'Error', 'Something was wrong!')
 
+
     def load_configures(self):
         try:
             # 载入配置文件
@@ -51,9 +55,10 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
                 self.Config = Parameters.Parameters(config_dic)
                 self.Config.adjust_parameters()
                 QtWidgets.QMessageBox.information(self, 'Success', 'You have loaded your configure file!')
-                self.enable_buttons(True)
+                self.enable_buttons(1)
         except:
             QtWidgets.QMessageBox.warning(self, 'Error', 'Something was wrong!')
+
 
     def delete_model(self):
         reply = QtWidgets.QMessageBox.warning(self, 'Delete Model', 'Delete the saved model?',
@@ -69,10 +74,11 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
             except:
                 QtWidgets.QMessageBox.warning(self, 'Error', 'Something was wrong!')
 
+
     def ShowFileDialog(self, tag):
         if tag == 1:
             try:
-                fname = QtWidgets.QFileDialog.getExistingDirectory()
+                fname = QtWidgets.QFileDialog.getExistingDirectory(directory='datasets')
                 if fname:
                     self.ImageSetPath.setText(fname)
                     Parameters.INPUT_DATA = fname
@@ -81,12 +87,13 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
                 QtWidgets.QMessageBox.warning(self, 'Error', 'Something was wrong!')
         elif tag == 2:
             try:
-                fname = QtWidgets.QFileDialog.getOpenFileName()
+                fname = QtWidgets.QFileDialog.getOpenFileName(directory='datasets')
                 if fname[0]:
                     self.ImagePath.setText(fname[0])
                     self.PredictResBrowser.setText("")
             except:
                 QtWidgets.QMessageBox.warning(self, 'Error', 'Something was wrong!')
+
 
     def learning_status_init(self):
         status_info = "Validation Percentage: %d\nTestSet Percentage: %d\n" \
@@ -95,12 +102,14 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
                          Parameters.LearningSteps, Parameters.BatchSize)
         self.TrainingStatusBrowser.setText(status_info)
 
+
     def update_statusinfo(self, status_info):
         self.TrainingStatusBrowser.append(status_info)
         self.TrainingStatusBrowser.moveCursor(QtGui.QTextCursor.End)
         # old_info = self.TrainingStatusBrowser.toPlainText()
         # new_info = status_info + '\n'+ old_info
         # self.TrainingStatusBrowser.setText(new_info)
+
 
     def update_predict(self, res):
         if res < 0:
@@ -109,8 +118,10 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
         else:
             self.PredictResBrowser.append("This image belongs to '%s'.\n" % (Parameters.LABEL_NAME_LIST[res]))
 
+
     def update_processBar(self, percentage):
         self.progressBar.setValue(percentage)
+
 
     def save_results(self):
         res_info = self.TrainingStatusBrowser.toPlainText()
@@ -122,6 +133,7 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
                 res_save.close()
         except:
             QtWidgets.QMessageBox.warning(self, 'Error', 'Something was wrong!')
+
 
     def disable_buttons(self, tag=1):
         if tag == 2:
@@ -136,17 +148,25 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
         self.WrongButton.setDisabled(True)
         self.RightButton.setDisabled(True)
 
+
     def enable_buttons(self, tag):
-        self.ImageSetBrowseButton.setEnabled(tag)
-        self.SettingsButton.setEnabled(tag)
-        self.StartTrainingButton.setEnabled(tag)
-        self.DelModelButton.setEnabled(tag)
-        self.SaveResButton.setEnabled(tag)
-        self.ImageBrowseButton.setEnabled(tag)
-        self.StartPredictButton.setEnabled(tag)
-        self.WrongButton.setEnabled(tag)
-        self.RightButton.setEnabled(tag)
-        self.LoadConfigButton.setEnabled(tag)
+        #self.detect_thread_status()
+        self.ImageSetBrowseButton.setEnabled(True)
+        self.SettingsButton.setEnabled(True)
+        self.StartTrainingButton.setEnabled(True)
+        self.DelModelButton.setEnabled(True)
+        self.SaveResButton.setEnabled(True)
+        self.ImageBrowseButton.setEnabled(True)
+        self.StartPredictButton.setEnabled(True)
+        self.LoadConfigButton.setEnabled(True)
+
+        if tag == 2:
+            self.WrongButton.setEnabled(True)
+            self.RightButton.setEnabled(True)
+
+
+        #self.restart_threads()
+
 
     def recovery_gui(self, tag):
         if tag == 1:
@@ -156,7 +176,8 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
             self.PredictResBrowser.setText("")
             self.ImagePath.setText("")
         self.progressBar.setValue(0)
-        self.enable_buttons(True)
+        self.enable_buttons(1)
+
 
     def error_process(self, error_key):
         if error_key == 1:
@@ -181,6 +202,7 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
             QtWidgets.QMessageBox.warning(self, 'Error', "Can not get the image bottleneck, please check your input.\n")
             self.recovery_gui(1)
 
+
     def wrong_prediction(self, feed_backs):
         reply = QtWidgets.QMessageBox.question(self, 'Oops',
                                                'Oops...Sorry for a bad prediction.\nDo you want to make a feedback?',
@@ -193,30 +215,34 @@ class MainGUI(Ui_MainWindow, QtWidgets.QMainWindow):
         dest_path = os.path.join(Parameters.TRAININGDATABASE, Parameters.MODEL_SAVE_NAME, str(feed_backs[0]))
         shutil.copy(feed_backs[1], dest_path)
         QtWidgets.QMessageBox.information(self, 'Nice', 'A nice prediction. So happy!')
-        os.remove(feed_backs[1])
+        if os.path.exists(feed_backs[1]):
+            os.remove(feed_backs[1])
+
 
     def handle_feed_back(self, feed_back_list):
-        feed_backs = feed_back_list
-        self.WrongButton.clicked.connect(lambda: self.wrong_prediction(feed_backs))
-        self.RightButton.clicked.connect(lambda: self.right_prediction(feed_backs))
+        self.PredictThread.stop()
+        while not self.PredictThread.isFinished():
+            self.PredictThread.wait()
+        self.feed_backs = feed_back_list
+
 
     def start_training(self):
         self.disable_buttons(tag=2)
         self.TrainThread = WorkThread.TrainingProcess()
         self.TrainThread.status_info.connect(self.update_statusinfo)
         self.TrainThread.done_percentage.connect(self.update_processBar)
-        self.TrainThread.finish_signal.connect(self.enable_buttons)
+        self.TrainThread.finish_signal.connect(lambda: self.enable_buttons(1))
         self.TrainThread.error_signal.connect(self.error_process)
         self.TrainThread.start()
 
     def picture_predict(self):
         self.disable_buttons(tag=2)
         self.PredictResBrowser.setText("Start predicting, please wait a moment...\n")
+        Parameters.input_image_path = self.ImagePath.text()
         try:
-            image_Path = self.ImagePath.text()
-            self.PredictThread = WorkThread.PicturePredict(image_Path)
+            self.PredictThread = WorkThread.PicturePredict()
             self.PredictThread.trigger.connect(self.update_predict)
-            self.PredictThread.finish_signal.connect(self.enable_buttons)
+            self.PredictThread.finish_signal.connect(lambda: self.enable_buttons(2))
             self.PredictThread.error_signal.connect(self.error_process)
             self.PredictThread.feed_back.connect(self.handle_feed_back)
             self.PredictThread.start()
